+++ src/pages/MainPage.tsx
@@ -1,18 +1,13 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Header } from '../components/Header';
 import { Footer } from '../components/Footer';
 import { CardsList } from '../components/CardsList';
-import { ArtGalleryService } from '../services/ArtGalleryService';
-import { API_URL, CARDS_COUNT_PER_PAGE } from '../constants';
+import { CARDS_COUNT_PER_PAGE } from '../constants';
 import { ICardData } from '../interfaces/ICardData';
-import { convertArtGalleryResponseToCards } from '../utils/apiDataConverter';
-import { IArtGalleryResponseSearch } from '../interfaces/IArtGalleryResponse';
-import {
-  getQueryFormLocalStorage,
-  saveQueryToLocalStorage,
-} from '../utils/querySaveTools';
+import { getQueryFormLocalStorage } from '../utils/querySaveTools';
 import { useSearchParams } from 'react-router-dom';
 import { IPaginatedArray } from '../interfaces/IPaginatedArray';
+import { loadCardsData } from '../utils/loadCardsData';
 
 export function MainPage() {
   const [data, setData] = useState<IPaginatedArray<ICardData> | null>(null);
@@ -24,50 +19,6 @@ export function MainPage() {
   const searchString = searchParams.get('search');
   const searchPageQuery = searchParams.get('page');
   const searchCount = searchParams.get('count');
-
-  const loadCardsData = useCallback(
-    async (
-      query: string | null = null,
-      page: number = 1,
-      cardsPerPageCount: number = CARDS_COUNT_PER_PAGE
-    ) => {
-      const artGalleryService = new ArtGalleryService(API_URL);
-
-      let response: IArtGalleryResponseSearch | null = null;
-
-      let pagesParam = {};
-      let cardsPerPageParam = {};
-      let queryParam = {};
-
-      if (page > 1) {
-        pagesParam = { page: page.toString() };
-      }
-
-      if (cardsPerPageCount !== CARDS_COUNT_PER_PAGE) {
-        cardsPerPageParam = { count: cardsPerPageCount.toString() };
-      }
-
-      if (query === null || query === '') {
-        response = await artGalleryService.getAll(page, cardsPerPageCount);
-        saveQueryToLocalStorage(null);
-      } else {
-        saveQueryToLocalStorage(query);
-        queryParam = {
-          search: query,
-        };
-        response = await artGalleryService.getByQueryString(
-          query,
-          page,
-          cardsPerPageCount
-        );
-      }
-      return {
-        cards: convertArtGalleryResponseToCards(response),
-        searchParams: { ...queryParam, ...pagesParam, ...cardsPerPageParam },
-      };
-    },
-    []
-  );
 
   const searchInArtGallery = async (query: string) => {
     setData(null);
@@ -108,7 +59,7 @@ export function MainPage() {
     loadCardsData(query, page, count)
       .then((data) => setData(data.cards))
       .catch((e) => console.error(e));
-  }, [loadCardsData, searchString, searchPageQuery, searchCount, cardsPerPage]);
+  }, [searchString, searchPageQuery, searchCount, cardsPerPage]);
 
   const handleSearchQueryChange = (query: string) => {
     if (query === searchQuery) {
